Replace float with double in several structs.

+++ GbxRemote.Net/Structs/PlayerNetInfo.cs
@@ -3,6 +3,6 @@ namespace GbxRemoteNet.Structs {
         public string IPAddress;
         public int StateUpdateLatency;
         public int LatestNetworkActivity;
-        public float PacketLossRate;
+        public double PacketLossRate;
     }
 }
